+++ src/main/java/com/kh/iMMUTABLE/repository/OrderRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
-    Order findByOrderId(int OrderId);
+    Order findByOrderId(long OrderId);
 
     List<Order> findByOrderStatus(OrderStatus OrderStatus);
